% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{category_league_settings}
\alias{category_league_settings}
\title{category_league_settings}
\usage{
category_league_settings(r_parsed)
}
\arguments{
\item{r_parsed}{variable created within y_league_settings by y_parse_response}
}
\description{
helper function called by y_league_settings()
}
\keyword{internal}
