% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.col_name_change_fn}
\alias{.col_name_change_fn}
\title{Prefix column names}
\usage{
.col_name_change_fn(x, y)
}
\arguments{
\item{x}{String to detect in column names. Passed to `names()`.}

\item{y}{Prefix string}
}
\value{
a sting
}
\description{
Used internally for prefixing resource and subresource names to column names.
}
\keyword{internals}
