% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.date_format_check}
\alias{.date_format_check}
\title{Check argument is date format}
\usage{
.date_format_check(x)
}
\arguments{
\item{...}{What to check.}
}
\description{
Checks the argument for "%Y-%m-%d" date format.
}
\keyword{internal}
