% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.date_param_fn}
\alias{.date_param_fn}
\title{Date parameter assignment.}
\usage{
.date_param_fn(game_key, game_date)
}
\arguments{
\item{game_key}{game key.}

\item{game_date}{game_date argument provided to YFAR function.}
}
\description{
Check that the date argument provided as argument to YFAR functions
is compatible with league.  In other words you can't supply a week integer
argument to daily leagues and you can't supply a date argument to weekly nfl leagues.
}
\details{
This function is for uri construction.
}
\keyword{internal}
