% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.draft_analysis_func}
\alias{.draft_analysis_func}
\title{Parse draft analysis sub-resource}
\usage{
.draft_analysis_func(x)
}
\arguments{
\item{x}{List containing a draft analysis element.}
}
\description{
Parse draft analysis sub-resource
}
\keyword{internal}
