% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.game_key_assign_fn}
\alias{.game_key_assign_fn}
\title{Assign game key from a vector of keys.}
\usage{
.game_key_assign_fn(key)
}
\arguments{
\item{key}{A vector of league game or player keys.}
}
\value{
A string
}
\description{
Assign game key from a vector of keys.
}
\keyword{internal}
