% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.game_meta_parse_fn}
\alias{.game_meta_parse_fn}
\title{Parse meta from game resource}
\usage{
.game_meta_parse_fn(x)
}
\arguments{
\item{x}{Game resource}
}
\description{
Parse meta from game resource
}
\keyword{internal}
