% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.game_resource_parse_fn}
\alias{.game_resource_parse_fn}
\title{Parse game resource}
\usage{
.game_resource_parse_fn(x, ...)
}
\arguments{
\item{x}{Games resource.}

\item{...}{Additional parse function passed to purrr::map_df.}
}
\description{
This function parses the games resource.
}
\details{
The ... argument allows passing an addition parse function to parse addition
sub-resources.
}
\keyword{internal}
