% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.game_stats_category_parse_fn}
\alias{.game_stats_category_parse_fn}
\title{Parse game stats data.}
\usage{
.game_stats_category_parse_fn(x)
}
\arguments{
\item{x}{Stats sub-resource of games resource}
}
\description{
Parse game stats data.
}
\keyword{internal}
