% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.id_check}
\alias{.id_check}
\title{id check}
\usage{
.id_check(id)
}
\arguments{
\item{id}{league_id supplied to y_function}
}
\description{
detect if league or team id provided to id argument
}
\keyword{internal}
