% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.key_resource_assign}
\alias{.key_resource_assign}
\title{Key check}
\usage{
.key_resource_assign(key)
}
\arguments{
\item{key}{league_key, team_key, player_key or game_key supplied to y_function}
}
\description{
Detect which key provided to key argument.
}
\keyword{internal}
