% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.league_key_check}
\alias{.league_key_check}
\title{League key check.}
\usage{
.league_key_check(key)
}
\arguments{
\item{key}{league key to check.}
}
\description{
Check format of league keys.
}
\keyword{internal}
