% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.league_meta_parse_fn}
\alias{.league_meta_parse_fn}
\title{Parse meta from league resource}
\usage{
.league_meta_parse_fn(x)
}
\arguments{
\item{x}{League resource}
}
\description{
Parse meta from league resource
}
\keyword{internal}
