% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.league_resource_fn}
\alias{.league_resource_fn}
\title{Parse return from league resource}
\usage{
.league_resource_fn(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_rosters.  this function calls .team_parse_fn helper.
}
\keyword{internal}
