% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.league_resource_parse_fn}
\alias{.league_resource_parse_fn}
\title{Parse return from league resource}
\usage{
.league_resource_parse_fn(x, pluck_args = NULL, fn)
}
\arguments{
\item{x}{Leagues resource.}

\item{pluck_args}{List of arguments passed to purrr::pluck.}

\item{fn}{Function to run on subresource.}
}
\value{
A tibble.
}
\description{
This function parses the league resource.
}
\details{
The ... argument allows passing an addition parse function to parse addition
sub-resources.

Sub-resource attached to team could be draft_results, roster, team_stats,
standings or match-ups.
}
\keyword{internal}
