% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.league_settings_parse_fn}
\alias{.league_settings_parse_fn}
\title{Parse league settings.}
\usage{
.league_settings_parse_fn(x)
}
\arguments{
\item{x}{Settings sub-resource.}
}
\description{
Parse league settings.
}
\keyword{internal}
