% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.multiple_resource_key_check}
\alias{.multiple_resource_key_check}
\title{Check key argument for homogeneity}
\usage{
.multiple_resource_key_check(
  key,
  e_key_types = c("games", "leagues", "teams", "players")
)
}
\arguments{
\item{key}{Vector of keys}

\item{e_key_types}{Vector of eligible key types, default is c("games", "leagues", "teams", "players").}
}
\value{
A list
}
\description{
This function checks the provided keys for uniformity.  If more that one type of key is detected
the function will select the key type appearing most often and removes the rest.
}
\details{
This function assigns resource variable.
This function alters key argument.
This function calls `.key_resource_assign()`
}
\keyword{internal}
