% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_id_check}
\alias{.player_id_check}
\title{player id check}
\usage{
.player_id_check(player_id)
}
\arguments{
\item{player_id}{league_id supplied to y_function}
}
\description{
Checks for presence of supplied player id and validity
}
\keyword{internal}
