% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.player_key_check}
\alias{.player_key_check}
\title{Player key check}
\usage{
.player_key_check(key)
}
\arguments{
\item{key}{player_key to check}
}
\description{
Check format of player_keys
}
\keyword{internal}
