% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_meta_func}
\alias{.player_meta_func}
\title{Parse player meta data element}
\usage{
.player_meta_func(x, ...)
}
\arguments{
\item{x}{Parsed content from response object}

\item{...}{Arguments passed on to purrr::pluck}
}
\description{
helper function called by y_functions to parse team meta data
}
\keyword{internal}
