% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_meta_parse_fn}
\alias{.player_meta_parse_fn}
\title{Parse meta data element from player resource.}
\usage{
.player_meta_parse_fn(x)
}
\arguments{
\item{x}{Player resource}
}
\description{
Parse meta data element from player resource.
}
\keyword{internal}
