% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_parse_fn}
\alias{.player_parse_fn}
\title{Parse return from player resource}
\usage{
.player_parse_fn(x)
}
\arguments{
\item{x}{list passed on from .roster_resource_fn}
}
\description{
helper function called by y_rosters.  This function calls .player_parse_fn helper.
}
\keyword{internal}
