% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_resource_parse_fn}
\alias{.player_resource_parse_fn}
\title{Parse player resource.}
\usage{
.player_resource_parse_fn(x, ...)
}
\arguments{
\item{x}{Player resource.}

\item{...}{Additional parse function passed to purrr::map_df.}
}
\description{
Parse player resource.
}
\keyword{internal}
