% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_stats_parse}
\alias{.player_stats_parse}
\title{Parse player stats}
\usage{
.player_stats_parse(x)
}
\arguments{
\item{x}{element to parse}
}
\value{
a tibble
}
\description{
Parse the stats resource of a player collection
}
\keyword{internal}
