% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.roster_parse_fn}
\alias{.roster_parse_fn}
\title{parse return from roster resource}
\usage{
.roster_parse_fn(x, ...)
}
\arguments{
\item{x}{list passed on from .roster_resource_fn}

\item{...}{arguments passed on to purrr::pluck}
}
\description{
helper function called by y_rosters.  this function calls .player_parse_fn helper.
}
\keyword{internal}
