% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.roster_resource_parse_fn}
\alias{.roster_resource_parse_fn}
\title{Parse roster resource.}
\usage{
.roster_resource_parse_fn(x, pluck_args, fn)
}
\arguments{
\item{x}{Roster resource.}

\item{pluck_args}{List of arguments passed to purrr::pluck.}

\item{fn}{Function to run on sub-resource.}
}
\value{
A tibble.
}
\description{
This function parses the rosters resource.
}
\details{
Calls .player_resource_parse_fn because rosters have players.
}
\keyword{internal}
