% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.seq_pages_fn}
\alias{.seq_pages_fn}
\title{Sequence uri page start and count params}
\usage{
.seq_pages_fn(start, count, i = 25)
}
\arguments{
\item{start}{What count to start at.}

\item{count}{Number of players to return.}

\item{i}{Response length.}
}
\value{
A vector of strings.
}
\description{
The Yahoo! API only returns 25 players at a time.  So if you want more than 25 players you
need to create a sequence of uri that call 25 players each.
}
\details{
I.E. If 50 players are desired you need 2 params:
`start=0;count=25` and `start=25;count=25`

This function takes a desired number of players (count) and how deep in the player
list to start (start) and turns them into a sequence of params with length 25.
}
\keyword{internal}
