% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.settings_subresource_parse_fn}
\alias{.settings_subresource_parse_fn}
\title{Parse settings sub-resource}
\usage{
.settings_subresource_parse_fn(x, ...)
}
\arguments{
\item{x}{List containing a matchup element.}

\item{...}{Takes a function and is passed on to do.call.}
}
\value{
A tibble
}
\description{
This function parses the settings sub-resource.
}
\details{
Called by `y_league_settings()`.
}
\keyword{internal}
