% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.standings_parse_fn}
\alias{.standings_parse_fn}
\title{Parse standings sub-resource}
\usage{
.standings_parse_fn(x)
}
\arguments{
\item{x}{List element containing standings resource.}
}
\description{
Parse standings sub-resource
}
\keyword{internal}
