% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.stats_data_func}
\alias{.stats_data_func}
\title{Parse stats data}
\usage{
.stats_data_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_matchups to parse team standings response
}
\keyword{internal}
