% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_id_check}
\alias{.team_id_check}
\title{team_id check}
\usage{
.team_id_check(team_id)
}
\arguments{
\item{team_id}{team id supplied to y_function}
}
\description{
Checks for presence of supplied team_id and validity
}
\keyword{internal}
