% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_info_func}
\alias{.team_info_func}
\title{team_info_func}
\usage{
.team_info_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_matchups to parse team_info from matchups response
}
\keyword{internal}
