% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_meta_func}
\alias{.team_meta_func}
\title{team_meta_data}
\usage{
.team_meta_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_ functions to parse team meta data
}
\keyword{internal}
