% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_meta_parse_fn}
\alias{.team_meta_parse_fn}
\title{Parse team meta data element}
\usage{
.team_meta_parse_fn(x)
}
\arguments{
\item{x}{Team resource}
}
\description{
Parse team meta data element
}
\keyword{internal}
