% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_parse_fn}
\alias{.team_parse_fn}
\title{parse return from team resource}
\usage{
.team_parse_fn(x)
}
\arguments{
\item{x}{list passed on from .league_resource_fn}
}
\description{
helper function called by y_rosters.  this function calls .roster_parse_fn helper.
}
\keyword{internal}
