% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_resource_parse_fn}
\alias{.team_resource_parse_fn}
\title{Parse return from team resource}
\usage{
.team_resource_parse_fn(x, ...)
}
\arguments{
\item{x}{Team resource.}

\item{...}{Additional parse function passed to purrr::map_df.}
}
\description{
This function parses the teams resource as well as the attached teams sub-resource.
}
\details{
The ... argument allows passing an addition parse function to parse addition
sub-resources.
}
\keyword{internal}
