% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.team_stats_parse_fn}
\alias{.team_stats_parse_fn}
\title{Parse team stats data.}
\usage{
.team_stats_parse_fn(x)
}
\arguments{
\item{x}{Stats sub-resource of teams resource}
}
\description{
Parse team stats data.
}
\keyword{internal}
