% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.token_check}
\alias{.token_check}
\title{Token Check}
\usage{
.token_check(token_name, api_token, ...)
}
\arguments{
\item{token_name}{assigned object name used when creating token with y_create_token().}

\item{api_token}{assigned value of token name to standardize token names within functions.}

\item{...}{argument passed onto ls()}
}
\description{
Checks supplied environment for a Token2.0 class.
Checks token_name argument is supplied.
Checks that the supplied token_name is class Token2.0.
}
\keyword{internal}
