% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.token_count}
\alias{.token_count}
\title{Token Count
Function called by .token_check and y_create_token()
Function called by .token_check
Check environment for a Token2.0 class.}
\usage{
.token_count(...)
}
\arguments{
\item{...}{environment name argument passed to ls()}
}
\description{
Token Count
Function called by .token_check and y_create_token()
Function called by .token_check
Check environment for a Token2.0 class.
}
\keyword{internal}
