% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.transaction_parse_fn}
\alias{.transaction_parse_fn}
\title{Parse transactions resource.}
\usage{
.transaction_parse_fn(x, pluck_args, fn)
}
\arguments{
\item{x}{element to parse}

\item{pluck_args}{List of arguments passed to purrr::pluck.}

\item{fn}{Function to run on subresource.}
}
\value{
a tibble
}
\description{
This function call .player_parse_fn
}
\keyword{internal}
