% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.transactions_func}
\alias{.transactions_func}
\title{transactions_func}
\usage{
.transactions_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_transactions to parse transaction response
}
\keyword{internal}
