% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.two_deep_subresource_parse}
\alias{.two_deep_subresource_parse}
\title{Parse a subresource that list a simple list of 1 deep lists.}
\usage{
.two_deep_subresource_parse(x)
}
\arguments{
\item{x}{Subresource containing lists of depth 1.}
}
\value{
A tibble
}
\description{
Function takes the name of the list and prefixes it any columns that
don't start with that prefix.
}
\keyword{internal}
