% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unlist_and_bind_fn}
\alias{.unlist_and_bind_fn}
\title{Unlist an element and bind.}
\usage{
.unlist_and_bind_fn(x, ...)
}
\arguments{
\item{x}{Object to unlist and bind.}

\item{...}{Arguments passed onto unlist2.}
}
\value{
A tibble.
}
\description{
This function combines `YFAR::unlist2` and `dplyr::bind_rows`.
}
\keyword{internal}
