% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.uri_gen_func}
\alias{.uri_gen_func}
\title{Generate uris.  Arugments are passed on to `httr::modify_url()`.}
\usage{
.uri_gen_func(
  resp_len = 25,
  league_id,
  resource,
  subresource,
  start,
  number_of_players,
  ...
)
}
\arguments{
\item{resp_len}{length of each response with a max of 25}

\item{league_id}{league_id as a string in the form "000.l.0000".  These ids can be found with `y_games()` and `y_teams()`.}

\item{resource}{resource from the api to be called}

\item{subresource}{subresource from the api to be called, takes a vector if more than one subresource}

\item{start}{Return start number}

\item{number_of_players}{Count of players to return}

\item{...}{api filter parameters}
}
\value{
a vector of strings
}
\description{
The Yahoo! Fantasy API will return a max of 25 players per call to the API.
This function divided by 25 and finds the remainder to construct the uri's
which are then passed to `y_get_response()`.
}
