% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.week_check}
\alias{.week_check}
\title{week check}
\usage{
.week_check(week)
}
\arguments{
\item{week}{team id supplied to y_function}
}
\description{
Checks for presence of supplied week and validity
}
\keyword{internal}
