% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.week_format_check}
\alias{.week_format_check}
\title{Check argument is week format}
\usage{
.week_format_check(x)
}
\arguments{
\item{x}{What to check.}
}
\description{
Checks the argument is an integer between 1 and 18.
}
\keyword{internal}
