% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.week_info_func}
\alias{.week_info_func}
\title{week_info_func}
\usage{
.week_info_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_matchups to parse week_info from matchups response
}
\keyword{internal}
