% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.week_stats_func}
\alias{.week_stats_func}
\title{week_stats_func}
\usage{
.week_stats_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_matchups to parse week stats from matchups response
}
\keyword{internal}
