% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.y_games_parse}
\alias{.y_games_parse}
\title{Parse games resource}
\usage{
.y_games_parse(x)
}
\arguments{
\item{x}{input list}
}
\value{
a tibble
}
\description{
Parse games resource
}
\keyword{internal}
