% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.y_get_response}
\alias{.y_get_response}
\title{y_get_response}
\usage{
.y_get_response(uri = NULL, token_name = NULL)
}
\arguments{
\item{uri}{URI being queried.}

\item{token_name}{Oauth token value assign by `y_create_token()`.}
}
\description{
Send GET request to YAHOO! api
}
\keyword{internal}
