% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.y_token_validate}
\alias{.y_token_validate}
\title{y_validate_token}
\usage{
.y_token_validate(token_name)
}
\arguments{
\item{token_name}{name used when assigning value of `y_create_token()`}
}
\value{
a string
}
\description{
y_validate_token
}
