% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.yahoo_hockey_stat_categories}
\alias{.yahoo_hockey_stat_categories}
\title{Create a tibble of the stat categories for yahoo fantasy hockey}
\usage{
.yahoo_hockey_stat_categories()
}
\value{
a tibble
}
\description{
Create a tibble of the stat categories for yahoo fantasy hockey
}
\keyword{internal}
