% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flattenable}
\alias{is_flattenable}
\title{Test if a list is flattenable.}
\usage{
is_flattenable(x)
}
\arguments{
\item{List}{to test.}
}
\value{
A boolean.
}
\description{
Usually combined with `rlang::flatten_if` or `YFAR::.unlist_and_bind`
}
\details{
This function is from `dplyr::bind_rows` so all credit to those cats and none
to me what-so-ever.  I hope this is placeholder credit until I learn how to cite this
properly.
}
\keyword{internal}
