% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pluckable}
\alias{is_pluckable}
\title{Test if a list is pluckable.}
\usage{
is_pluckable(x)
}
\arguments{
\item{Element}{to test}
}
\value{
A list.
}
\description{
Pluckability is defined as a list which contains a names list of length 1 which is flattenable.
}
\keyword{internal}
