% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{league_id_check}
\alias{league_id_check}
\title{league_id check}
\usage{
league_id_check(x)
}
\arguments{
\item{x}{league_id supplied to y function}
}
\description{
Checks given league_id for validity
}
\keyword{internal}
