% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_pre_process_fn}
\alias{list_pre_process_fn}
\title{Preprocess list}
\usage{
list_pre_process_fn(x)
}
\arguments{
\item{x}{List to preprocess}
}
\value{
A list
}
\description{
Convert elements to character using `as.character()`
Prune unwanted `count` eleements.
}
\keyword{internal}
