% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{player_slate_func}
\alias{player_slate_func}
\title{player_slate_func}
\usage{
player_slate_func(x, ...)
}
\arguments{
\item{x}{parsed content from response object}

\item{...}{arguments passed onto purrr::pluck}
}
\description{
helper function called by y_player_slate to parse player info from players response
}
\keyword{internal}
