% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stat_winner_func}
\alias{stat_winner_func}
\title{stats_winner_func}
\usage{
stat_winner_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_matchups to parse stat_winners from matchups response
}
\keyword{internal}
