% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stats_data_func}
\alias{stats_data_func}
\title{stats_data_func}
\usage{
stats_data_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_standings to parse team standings response
}
\keyword{internal}
