% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{team_stats_func}
\alias{team_stats_func}
\title{team_stats_func}
\usage{
team_stats_func(x)
}
\arguments{
\item{x}{parsed content from response object}
}
\description{
helper function called by y_team_stats to parse team stats response
}
\keyword{internal}
