% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist2}
\alias{unlist2}
\title{Unlist with janitor::make_clean_names.}
\usage{
unlist2(x, ...)
}
\arguments{
\item{x}{Object to unlist.}

\item{...}{Arguments passed to `base::unlist`.}
}
\value{
A vector.
}
\description{
This function is a wrapper around `base::unlist` where the names attributes are clean with janitor::make_clean_names.
}
\keyword{internal}
