% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_draft_adp.R
\name{y_draft_adp}
\alias{y_draft_adp}
\title{Get transaction draft ADP data from Yahoo! Fantasy API}
\usage{
y_draft_adp(key = NULL, token_name = NULL, debug = FALSE, quiet = TRUE)
}
\arguments{
\item{key}{A vector of game, league or player keys.}

\item{token_name}{Name used for assignment when creating token object with y_create_token().}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Suppress function print calls.}
}
\value{
A tibble
}
\description{
Yahoo drafts are accompanied by ADP's and this function returns them.
Values return include average_pick, average_round, average_cost and percent drafted.
}
\details{
The return from this function is similar to `y_player_slate()` but returns the ADP's as additional columns.

Function accepts a single game key, a single league key or a vector of player keys.

Function is memoised.
}
\examples{
# Not Run
# y_draft_adp(key = "411", token_name = my_token, debug = FALSE)
# y_draft_adp(key = vector_of_player_keys, token_name = my_token, debug = FALSE)
}
