% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_games.R
\name{y_games}
\alias{y_games}
\title{Get all Yahoo! Fantasy hockey meta data.}
\usage{
y_games()
}
\value{
tibble
}
\description{
Returns a tibble containing all the meta data for all fantasy hockey
games and leagues participated in by the logged in user including those which
are currently active.
}
