% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{y_get_response}
\alias{y_get_response}
\title{y_get_response}
\usage{
y_get_response(x, y)
}
\arguments{
\item{x}{league_id supplied to y function}

\item{y}{api_token value assign by y_create_token()}
}
\description{
Send GET request to YAHOO! api
}
\keyword{internal}
