% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_league_settings.R
\name{y_league_settings}
\alias{y_league_settings}
\title{Get league settings data from Yahoo! Fantasy API.}
\usage{
y_league_settings(
  league_key = NULL,
  token_name = NULL,
  debug = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{league_key}{League key as a string or vector of strings in the form "000.l.0000".
League key can be found with `y_games()`.}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble.
}
\description{
Each Yahoo! Fantasy league can have unique settings.  Given a valid league key this
  function returns those settings for you.
}
\details{
Function returns a row of data for each league key provided.

  Returns upwards of 60 columns of data with some nested tibbles.

  Learn more in vignette.
}
