% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_matchups.R
\name{y_matchups}
\alias{y_matchups}
\title{Get matchup stats and results data from Yahoo! Fantasy API}
\usage{
y_matchups(team_id = NULL, token_name = NULL, week = NULL)
}
\arguments{
\item{team_id}{as a string in the form "000.l.0000".  Team id can be found with y_teams().}

\item{token_name}{assigned object name used when creating token with y_create_token().}

\item{week}{the week of fantasy season to return. Default NULL will return all weeks of season.}
}
\value{
a list
}
\description{
y_matchups() gets matchup stats and results for a given league and week.
It returns a list with the meta data as well as tibble containing the matchups data
in the element `matchups_data`.
}
