% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{y_parse_response}
\alias{y_parse_response}
\title{y_parse_response}
\usage{
y_parse_response(x, ...)
}
\arguments{
\item{x}{league_id supplied to y function}

\item{...}{arguments passed onto purrr::pluck}
}
\description{
Parse response from y_get_response()
}
\keyword{internal}
