% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_player_stats.R
\name{y_player_stats}
\alias{y_player_stats}
\title{Get player stats data from Yahoo! Fantasy API.}
\usage{
y_player_stats(
  player_key = NULL,
  token_name = NULL,
  game_date = NULL,
  debug = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{player_key}{Vector of player keys. Key usually in the form xxx.p.xxxx.
Found with `y_players()`, `y_player_slate()` or `y_rosters()`}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{game_date}{Date or week of fantasy season.
Default NULL will return aggregate stats for current season.  Accepts a vector of dates.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble.
}
\description{
Want stats for a player or group of players?  Provide a vector of player keys and ye shall receive.
}
\details{
Want stats for a specific date? Provide a vector of game dates in form YYYY-MM-DD (%Y-%m-%d).
  Default game date NULL will return aggregate stats for the season.
  For weekly sports such as nfl you can provide an integer denoting a fantasy week.
}
