% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_players.R
\name{y_players}
\alias{y_players}
\title{Get player data from Yahoo! Fantasy API}
\usage{
y_players(
  league_id = NULL,
  token_name = NULL,
  start = 0,
  number_of_players = 100,
  debug = FALSE,
  ...
)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}

\item{start}{Where to start the uris i.e. 50}

\item{number_of_players}{Number of players}

\item{debug}{Returns a list of data uri call and content.  Useful for debugging.}

\item{...}{URI filter arguments (as characters) passed onto internal .uri_gen_func.}
}
\value{
a list or tibble
}
\description{
This function is not intended to get all players in a league.  To get all players in your league use the
function `y_player_slate()`.
}
\details{
Function returns a tibble containing info on all players in your leagues universe.
By default the players are sorted by "Actual Rank"

This function is intented to get a subset of players i.e. top 100 players sorted by AR (actual rank).
If you want all players in that exist in your leagues universe use `y_player_slate()`
}
