% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_scoreboard.R
\name{y_scoreboard}
\alias{y_scoreboard}
\title{Get player data from Yahoo! Fantasy API}
\usage{
y_scoreboard(league_id, token_name)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Function returns a tibble containing info on all players in your leagues universe.
By default the players are sorts by "Actual Rank"
}
