% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_standings.R
\name{y_standings}
\alias{y_standings}
\title{Get standings data from Yahoo! Fantasy API.}
\usage{
y_standings(team_key = NULL, token_name = NULL, debug = FALSE, quiet = TRUE)
}
\arguments{
\item{team_key}{A vector of team keys as a string in the form 000.l.0000.t.0".}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble.
}
\description{
Supply a team key or vector of team keys to return standings data.
This function is best use with y_teams to supply a vector to team keys from a league.
It seems you should be able to supply a league key to this function but for some reason the request returns
a bunch of team stats data and I wanted to keep that in the `y_team_stats()` function.
}
\details{
This function does not return individual category win data form H2H leagues.
}
