% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_stats_categories.R
\name{y_stats_categories}
\alias{y_stats_categories}
\title{Get all existing stats categories for a game from the Yahoo! Fantasy API.}
\usage{
y_stats_categories(
  game_key = NULL,
  token_name = NULL,
  debug = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{game_key}{2 or 3 digit number of the game your league is playing.  Found with `y_games()`.}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{debug}{Returns a list of data uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A nested tibble with a game_key column and nested stats column.
}
\description{
Note not all stats that exist in a game will be used by your league.
}
