% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_team_stats.R
\name{y_team_stats}
\alias{y_team_stats}
\title{Get team scoring stats from Yahoo! Fantasy API.}
\usage{
y_team_stats(
  key = NULL,
  token_name = NULL,
  week = NULL,
  debug = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{key}{League key or team key as a string in the form "000.l.0000" or "000.l.0000.t.0".}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{week}{Week of fantasy season to return.
-Accepts 3 arguments:
  - `current` returns current week of season.
  - An integer corresponding to a week of season.
  - If NULL will return aggregated season stats.}

\item{debug}{returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble
}
\description{
Returns team or league stats depending on what type of key is
  provided as an argument to key.
}
