% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_teams.R
\name{y_teams}
\alias{y_teams}
\title{Get team data of a given Yahoo! Fantasy league.}
\usage{
y_teams(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package] with both nested and un-nested columns.
}
\description{
Teams in a league make moves, have co-managers, have a waiver prioity etc.  This function
returns all team level meta data from the Yahoo! Fantasy API.
}
