% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_weeks.R
\name{y_weeks}
\alias{y_weeks}
\title{Title}
\usage{
y_weeks(game_id = NULL, token_name = NULL)
}
\arguments{
\item{game_id}{a 3 or 4 digit integer.  It is the first 3 digits of league_id.
Can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a list
}
\description{
Title
}
