% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.game_resource_parse_fn}
\alias{.game_resource_parse_fn}
\title{Parse game resource}
\usage{
.game_resource_parse_fn(x, pluck_args = NULL, fn)
}
\arguments{
\item{x}{Games resource.}

\item{pluck_args}{List of arguments passed to purrr::pluck.}

\item{fn}{Function to run on sub-resource.}
}
\value{
A tibble.
}
\description{
This function parses the games resource.
}
\details{
The ... argument allows passing an addition parse function to parse addition
sub-resources.
}
\keyword{internal}
