% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.league_resource_parse_fn}
\alias{.league_resource_parse_fn}
\title{Parse return from league resource}
\usage{
.league_resource_parse_fn(x, ...)
}
\arguments{
\item{x}{Leagues resource.}

\item{...}{Additional parse function passed to purrr::map_df.}
}
\description{
This function parses the league resource.
}
\details{
The ... argument allows passing an addition parse function to parse addition
sub-resources.

Sub-resource attached to team could be draft_results, roster, team_stats,
standings or match-ups.
}
\keyword{internal}
