% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.matchup_parse_fn}
\alias{.matchup_parse_fn}
\title{Parse matchup subresource}
\usage{
.matchup_parse_fn(x)
}
\arguments{
\item{x}{input}
}
\value{
a tibble or list
}
\description{
Called by y_matchups
}
\keyword{internal}
