% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.player_resource_parse_fn}
\alias{.player_resource_parse_fn}
\title{Parse player resource.}
\usage{
.player_resource_parse_fn(x, pluck_args = NULL, fn)
}
\arguments{
\item{x}{Player resource.}

\item{pluck_args}{List of arguments passed to purrr::pluck.}

\item{fn}{Function to run on subresource.}
}
\value{
A tibble.
}
\description{
Parse player resource.
}
\keyword{internal}
