% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.roster_resource_parse_fn}
\alias{.roster_resource_parse_fn}
\title{Parse roster resource.}
\usage{
.roster_resource_parse_fn(x)
}
\arguments{
\item{x}{Roster resource.}
}
\description{
This function parses the rosters resource.
}
\details{
Calls .player_resource_parse_fn because rosters have players.
}
\keyword{internal}
