% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.uri_gen_func}
\alias{.uri_gen_func}
\title{Generate uris.  Arugments are passed on to `httr::modify_url()`.}
\usage{
.uri_gen_func(
  number_of_players = 100,
  start = 0,
  resp_len = 25,
  resource,
  league_id,
  subresource,
  status = "ALL",
  sort = "OR"
)
}
\arguments{
\item{number_of_players}{number of players}

\item{start}{where to start the uris i.e. 50}

\item{resp_len}{length of each response with a max of 25}

\item{resource}{resource from the api to be called}

\item{league_id}{league_id as a string in the form "000.l.0000".  These ids can be found with `y_games()` and `y_teams()`.}

\item{subresource}{subresource from the api to be called, takes a vector if more than one subresource}

\item{status}{status of players to return ("ALL", "A", "FA", "T", "W", "K")}

\item{sort}{sort order of the players to return (stat_id, "OR", "AR", "PTS", "NAME(last, first)")}
}
\value{
a vector of strings
}
\description{
The Yahoo! Fantasy API will return a max of 25 players per call to the API.
This function divided by 25 and finds the remainder to construct the uri's
which are then passed to `y_get_response()`.
}
