% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flattenable}
\alias{is_flattenable}
\title{Test if a list is flattenable.}
\usage{
is_flattenable(x)
}
\arguments{
\item{List}{to test.}
}
\value{
A boolean.
}
\description{
Usually combined with `rlang::flatten_if` or `YFAR::.unlist_and_bind`
}
\keyword{internal}
