% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_create_token.R
\name{y_create_token}
\alias{y_create_token}
\title{Create an Oauth2.0 token}
\usage{
y_create_token(my_key = NULL, my_secret = NULL, app_name = NULL)
}
\arguments{
\item{my_key}{a string provided my registering an application with Yahoo Developer Network}

\item{my_secret}{a string provided my registering an application with Yahoo Developer Network}

\item{app_name}{a string name of app, not necessarily what you provided to Yahoo Developer Network}
}
\value{
Token2.0 object.
}
\description{
Creates an Oauth2.0 token object used to access the Yahoo Fantasy API
See httr::Token2.0 for details and functionality.
}
