% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_draft_results.R
\name{y_draft_results}
\alias{y_draft_results}
\title{Get draft results}
\usage{
y_draft_results(id = NULL, token_name = NULL, debug = FALSE)
}
\arguments{
\item{id}{League id or team id as a string in the form "000.l.0000" or "000.l.0000.t.0".  These ids can be found with `y_games()` and `y_teams()`.}

\item{token_name}{Assigned object name used when creating token with y_create_token().}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}
}
\value{
a list or tibble
}
\description{
Get draft results
}
