% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_draft_results.R
\name{y_draft_results}
\alias{y_draft_results}
\title{Get draft results}
\usage{
y_draft_results(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get draft results
}
