% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_games.R
\name{y_games}
\alias{y_games}
\title{Get all Yahoo! Fantasy hockey meta data.}
\usage{
y_games(token_name = NULL, debug = FALSE)
}
\arguments{
\item{token_name}{Assigned object name used when creating token with y_create_token().}

\item{debug}{returns a list of data such as uri call and content.  Useful for debugging.}
}
\value{
a tibble
}
\description{
Returns a tibble containing all the meta data for all fantasy hockey
games and leagues participated in by the logged in user including those which
are currently active.
}
\details{
Function is memoised
}
