% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_games.R
\name{y_games}
\alias{y_games}
\title{Get all Yahoo! Fantasy hockey meta data.}
\usage{
y_games(token_name = NULL)
}
\arguments{
\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Returns a tibble containing all the meta data for all fantasy hockey
games and leagues participated in by the logged in user including those which
are currently active.
}
