% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_games.R
\name{y_games}
\alias{y_games}
\title{Get user game data from Yahoo! Fantasy API}
\usage{
y_games(token_name = NULL, debug = FALSE, quiet = TRUE)
}
\arguments{
\item{token_name}{Name used for assignment when creating token object with y_create_token().}

\item{debug}{returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Toggle function messages.}
}
\value{
A tibble
}
\description{
Returns a tibble containing descriptive meta data for all Yahoo! fantasy
   games and leagues participated in by the logged in user.  Data returned
   describes both past and current games and includes but is not limited to
   game codes, year, scoring type and league name.
}
\details{
Function is memoised
}
