% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_matchups.R
\name{y_matchups}
\alias{y_matchups}
\title{Get match-up stats and results data from Yahoo! Fantasy API}
\usage{
y_matchups(
  team_key = NULL,
  token_name = NULL,
  week = NULL,
  debug = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{team_key}{A string in the form "000.l.0000.t.00".  Team key can be found with `y_teams()`.}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{week}{Week of fantasy season to return. Default NULL will return all past and current weeks of season.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble
}
\description{
Returns match-up stats and results for a given teams and weeks.
  Option to provide a week agrument which will return data for that week.
  Default NULL will return all past and current weeks of season.
  Vector of weeks also accepted.
}
\details{
Only past or current match-ups will be returned.
}
