% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_matchups.R
\name{y_matchups}
\alias{y_matchups}
\title{Get matchup stats and results data from Yahoo! Fantasy API}
\usage{
y_matchups(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package] with nested columns
}
\description{
y_matchups() gets matchup stats and results for a given league.
It returns a tibble of nested columns with a length equal to the number of teams
in the league.
}
