% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_player_slate.R
\name{y_player_slate}
\alias{y_player_slate}
\title{Get player data from Yahoo! Fantasy API}
\usage{
y_player_slate(
  league_id = NULL,
  token_name = NULL,
  start = 0,
  number_of_players = 100
)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}

\item{start}{Return start number}

\item{number_of_players}{Count of players to return}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Function returns a tibble containing info on all players in your leagues universe.
By default the players are sorts by "Actual Rank"
}
