% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_player_slate.R
\name{y_player_slate}
\alias{y_player_slate}
\title{Retrieve the full slate of players for provided league id}
\usage{
y_player_slate(league_id, token_name, debug = FALSE, ...)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  These ids can be found with `y_games()` and `y_teams()`.}

\item{token_name}{Assigned object name used when creating token with `y_create_token()`.}

\item{debug}{Print uri and page counts to console as functions runs.  Useful for debugging.}

\item{...}{URI filter arguments passed onto internal .uri_gen_func.}
}
\value{
A tibble
}
\description{
This function is not intended to get a subset of players, i.e. top 100 players.
Use `y_players()` for that case.
}
\details{
Note: this function uses janitor::make_clean_names and as a result is a bit slow.
}
