% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_player_stats.R
\name{y_player_stats}
\alias{y_player_stats}
\title{Get player stats}
\usage{
y_player_stats(
  players = NULL,
  token_name = NULL,
  game_date = NULL,
  debug = FALSE
)
}
\arguments{
\item{players}{player keys.  Keys are in the form xxx.p.xxxx and found using
`y_players()`, `y_player_slate()` or `y_rosters()`}

\item{token_name}{Assigned object name used when creating token with y_create_token().}

\item{game_date}{Date of fantasy season in form YYYY-MM-DD to return.
Default is NULL which will return aggregate stats for current season.  Accepts a vector of dates.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}
}
\value{
a tibble or list
}
\description{
Get player stats
}
