% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_players.R
\name{y_players}
\alias{y_players}
\title{Get player data from Yahoo! Fantasy API.}
\usage{
y_players(
  key = NULL,
  token_name = NULL,
  start = 0,
  number_of_players = 100,
  quiet = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{key}{Game or league key as a string in the form "000" or 000.l.0000".  Can be found with `y_games()`.}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{start}{Where to start the count in uris i.e. 50.  Default is 0.}

\item{number_of_players}{Number of players to return.  Default is 100}

\item{quiet}{Print function activity.}

\item{debug}{Returns a list of data uri call and content.  Useful for debugging.}

\item{...}{Other URI filter arguments.  See vignette for full list.}
}
\value{
A tibble
}
\description{
This function is not intended to get all players in a league.  To get all players in your league use the
function `y_player_slate()`.
}
\details{
This function is intended to get a subset of players i.e. top 100 players sorted by AR (actual rank).

Takes many filter and sort arguments via ...
E.G. sort = 1 will return however many players you requested sorted by stat 1.
See vignette for all possibilities.
}
