% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_rosters.R
\name{y_rosters}
\alias{y_rosters}
\title{Get roster data for a Yahoo Fantasy League.}
\usage{
y_rosters(id = NULL, token_name = NULL, date = NULL)
}
\arguments{
\item{id}{league id or team id as a string in the form "000.l.0000" or "000.l.0000.t.0".  These ids can be found with `y_games()` and `y_teams()`.}

\item{token_name}{assigned object name used when creating token with `y_create_token()`.}

\item{date}{date of fantasy season in form YYYY-MM-DD to return.  Default is null and will return current date.}
}
\value{
a list
}
\description{
Supply a team or league id to the function to retrieve roster data.  If a date
is supplied the roster data for that date will be returned.
}
