% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_standings.R
\name{y_standings}
\alias{y_standings}
\title{Get standings data from Yahoo! Fantasy API}
\usage{
y_standings(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a list
}
\description{
Get standings data, divison data playoff seed and team matchup results.
This function does not return individual category win data form H2H leagues.
}
