% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_stats_categories.R
\name{y_stats_categories}
\alias{y_stats_categories}
\title{Get all exisiting stats categories for a game.}
\usage{
y_stats_categories(game_key = 411, token_name = NULL)
}
\arguments{
\item{game_key}{3 digit number of the game your league is playing.  Found with `y_games()`.}

\item{token_name}{Assigned object name used when creating token with `y_create_token()`.}
}
\value{
a tibble
}
\description{
Get all exisiting stats categories for a game.
}
