% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_team_stats.R
\name{y_team_stats}
\alias{y_team_stats}
\title{Get team level stats for scoring categories}
\usage{
y_team_stats(id = NULL, token_name = NULL, week = "current")
}
\arguments{
\item{id}{league id or team id as a string in the form "000.l.0000" or "000.l.0000.t.0".  These ids can be found with `y_games()` and `y_teams()`.}

\item{token_name}{Assigned object name used when creating token with `y_create_token()`.}

\item{week}{week of fantasy season to return.  Accepts 3 arguments `current`, `NULL`, `a number`.
`current` by default returns current week, null returns aggregated season stats, number returns stats for that week.}
}
\value{
a list
}
\description{
returns team or league stats depending on what id (league or team) is supplied to `id`
}
