% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_team_stats.R
\name{y_team_stats}
\alias{y_team_stats}
\title{Get team level stats for scoring categories}
\usage{
y_team_stats(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a list
}
\description{
Get team level stats for scoring categories
}
