% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_teams.R
\name{y_teams}
\alias{y_teams}
\title{Get team level meta data from Yahoo! Fantasy API.}
\usage{
y_teams(league_key = NULL, token_name = NULL, debug = FALSE, quiet = TRUE)
}
\arguments{
\item{league_key}{League key as a string in the form "000.l.0000".  League id can be found with `y_games()`.}

\item{token_name}{Name used for assignment when creating token object with `y_create_token()`.}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble.
}
\description{
Teams in a league make moves, have co-managers, have a waiver prioity etc.  This function
returns all team level meta data from the Yahoo! Fantasy API.
}
