% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_transactions.R
\name{y_transactions}
\alias{y_transactions}
\title{Get transaction data from Yahoo! Fantasy API}
\usage{
y_transactions(league_id = NULL, token_name = NULL)
}
\arguments{
\item{league_id}{League id as a string in the form "000.l.0000".  League id can be found with y_games().}

\item{token_name}{Assigned object name used when creating token with y_create_token().}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get transaction data from Yahoo! Fantasy API
}
