% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_weeks.R
\name{y_weeks}
\alias{y_weeks}
\title{Get dates of fantasy weeks from Yahoo! Fantasy API.}
\usage{
y_weeks(game_key = NULL, token_name = NULL, debug = FALSE, quiet = TRUE)
}
\arguments{
\item{game_key}{A game key or a vector of game keys.  Can be found with y_games().}

\item{token_name}{Name used for assignment when creating token object with y_create_token().}

\item{debug}{Returns a list of data such as uri call and content.  Useful for debugging.}

\item{quiet}{Print function activity.}
}
\value{
A tibble
}
\description{
Want to see start and end dates for fantasy weeks of a particular league?  Curious how long
  long a fantasy week is?  If yes, this is your function.
}
